+++ packages/terra-list/CHANGELOG.md
@@ -3,6 +3,19 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Added documentation on required peerDependencies
+
+### Changed
+* Update tests to use `Terra.describeViewports` for setting viewports and use `Terra.it.validatesElement`
+
+4.10.0  - (July 11, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
+4.9.0 - (June 28, 2019)
+------------------
 ### Changed
 * Fixed accessibility issue in list item tests
 * Update mixin import to work better with codesandbox.io